+++ Cappuccino.Web/Areas/SystemManage/Controllers/SysDictDetailController.cs
@@ -27,14 +27,14 @@ namespace Cappuccino.Web.Areas.SystemManage.Controllers
 
         #region 视图        
         [HttpGet, CheckPermission("system.dict.create")]
-        public ActionResult Create(int dictId)
+        public ActionResult Create(long dictId)
         {
             ViewBag.DictId = dictId;
             return View();
         }
 
         [HttpGet, CheckPermission("system.dict.edit")]
-        public ActionResult Edit(int id)
+        public ActionResult Edit(long id)
         {
             SysDictDetailEntity viewModel = _sysDictDetailService.GetList(x => x.Id == id).FirstOrDefault();
             return View(viewModel.EntityMap());
@@ -84,7 +84,7 @@ namespace Cappuccino.Web.Areas.SystemManage.Controllers
 
         [HttpPost, CheckPermission("system.dict.delete")]
         [LogOperate(Title = "删除字典详情", BusinessType = (int)OperateType.Delete)]
-        public ActionResult Delete(int id)
+        public ActionResult Delete(long id)
         {
             try
             {
@@ -104,7 +104,7 @@ namespace Cappuccino.Web.Areas.SystemManage.Controllers
             try
             {
                 var idsArray = idsStr.Substring(0, idsStr.Length).Split(',');
-                int[] ids = Array.ConvertAll<string, int>(idsArray, int.Parse);
+                long[] ids = Array.ConvertAll<string, long>(idsArray, long.Parse);
                 var result = _sysDictDetailService.DeleteBy(x => ids.Contains(x.Id)) > 0 ? WriteSuccess("数据删除成功") : WriteError("数据删除失败");
                 return result;
             }
@@ -144,7 +144,7 @@ namespace Cappuccino.Web.Areas.SystemManage.Controllers
             return Json(Pager.Paging(list, totalCount), JsonRequestBehavior.AllowGet);
         }
 
-        public JsonResult GetMaxSortCode(int dictId)
+        public JsonResult GetMaxSortCode(long dictId)
         {
             int maxSortCode = _sysDictDetailService.GetMaxSortCode(dictId);
             var result = new { Status = 0, Message = "查询成功", Data = maxSortCode };
